[PLC] [Refac] Cleaned the core Type module up
