+++ language-plutus-core/common/Common.hs
@@ -13,14 +13,14 @@ module Common
     , nestedGoldenVsDocM
     ) where
 
-import           Language.PlutusCore.Pretty
+import           PlutusPrelude        hiding ((</>))
 
-import           Control.Monad.Reader       (Reader, runReader)
-import qualified Control.Monad.Reader       as Reader
-import qualified Data.ByteString.Lazy       as BSL
-import           Data.Text                  (Text)
-import           Data.Text.Encoding         (encodeUtf8)
-import           System.FilePath            ((</>))
+import           Control.Monad.Reader (Reader, runReader)
+import qualified Control.Monad.Reader as Reader
+import qualified Data.ByteString.Lazy as BSL
+import           Data.Text            (Text)
+import           Data.Text.Encoding   (encodeUtf8)
+import           System.FilePath      ((</>))
 import           Test.Tasty
 import           Test.Tasty.Golden
