+++ language-plutus-core/generators/Language/PlutusCore/Generators/Internal/TypedBuiltinGen.hs
@@ -18,12 +18,13 @@ module Language.PlutusCore.Generators.Internal.TypedBuiltinGen
     , genTypedBuiltinDivide
     ) where
 
-import           Language.PlutusCore.Constant
-import           Language.PlutusCore.Name
-import           Language.PlutusCore.Type
 import           PlutusPrelude
 
 import           Language.PlutusCore.Generators.Internal.Dependent
+
+import           Language.PlutusCore.Constant
+import           Language.PlutusCore.Name
+import           Language.PlutusCore.Type
 
 import qualified Data.ByteString.Lazy                              as BSL
 import           Data.Functor.Identity
