+++ language-plutus-core/generators/Language/PlutusCore/Generators/Internal/Utils.hs
@@ -17,7 +17,7 @@ module Language.PlutusCore.Generators.Internal.Utils
     , prettyPlcErrorString
     ) where
 
-import           Language.PlutusCore.Pretty
+import           Language.PlutusCore.Type
 
 import           Control.Monad.Morph
 import           Control.Monad.Reader
