+++ language-plutus-core/language-plutus-core.cabal
@@ -36,10 +36,10 @@ library
         Language.PlutusCore.Rename
         Language.PlutusCore.Normalize
         Language.PlutusCore.Normalize.Internal
-        Language.PlutusCore.Pretty
         Language.PlutusCore.View
         Language.PlutusCore.Subst
         Language.PlutusCore.Name
+        Language.PlutusCore.Type
         Language.PlutusCore.DeBruijn
         Language.PlutusCore.Check.Uniques
         Language.PlutusCore.FsTree
@@ -72,8 +72,16 @@ library
     build-tool-depends: alex:alex -any, happy:happy >=1.17.1
     hs-source-dirs: src prelude stdlib examples generators common
     other-modules:
-        Language.PlutusCore.Type
-        Language.PlutusCore.Lexer.Type
+        Language.PlutusCore.Pretty.ConfigName
+        Language.PlutusCore.Type.Core
+        Language.PlutusCore.Type.Plated
+        Language.PlutusCore.Type.Instance.Pretty.Common
+        Language.PlutusCore.Type.Instance.Pretty.Classic
+        Language.PlutusCore.Type.Instance.Pretty.Plc
+        Language.PlutusCore.Type.Instance.Pretty.Readable
+        Language.PlutusCore.Type.Instance.Pretty
+        Language.PlutusCore.Type.Instance.Recursive
+        Language.PlutusCore.Type.Instance
         Language.PlutusCore.Constant.Apply
         Language.PlutusCore.Constant.Dynamic.BuiltinName
         Language.PlutusCore.Constant.Dynamic.Call
@@ -85,9 +93,7 @@ library
         Language.PlutusCore.Constant.Make
         Language.PlutusCore.Constant.Name
         Language.PlutusCore.Constant.Typed
-        Language.PlutusCore.Pretty.Classic
-        Language.PlutusCore.Pretty.Plc
-        Language.PlutusCore.Pretty.Readable
+        Language.PlutusCore.Lexer.Type
         Language.PlutusCore.Error
         Language.PlutusCore.Size
         Language.PlutusCore.TypeCheck.Internal
