+++ language-plutus-core/src/Language/PlutusCore.hs
@@ -1,4 +1,3 @@
-
 module Language.PlutusCore
     (
       -- * Parser
@@ -97,10 +96,8 @@ module Language.PlutusCore
     , serialisedSize
     ) where
 
-import           Control.Monad.Except
-import qualified Data.ByteString.Lazy                     as BSL
-import qualified Data.Text                                as T
-import           Data.Text.Prettyprint.Doc
+import           PlutusPrelude
+
 import           Language.PlutusCore.CBOR                 ()
 import qualified Language.PlutusCore.Check.Normal         as Normal
 import qualified Language.PlutusCore.Check.Uniques        as Uniques
@@ -112,14 +109,16 @@ import           Language.PlutusCore.Lexer.Type
 import           Language.PlutusCore.Name
 import           Language.PlutusCore.Normalize
 import           Language.PlutusCore.Parser
-import           Language.PlutusCore.Pretty
 import           Language.PlutusCore.Quote
 import           Language.PlutusCore.Rename
 import           Language.PlutusCore.Size
 import           Language.PlutusCore.Type
 import           Language.PlutusCore.TypeCheck            as TypeCheck
 import           Language.PlutusCore.View
-import           PlutusPrelude
+
+import           Control.Monad.Except
+import qualified Data.ByteString.Lazy                     as BSL
+import qualified Data.Text                                as T
 
 -- | Given a file at @fibonacci.plc@, @fileType "fibonacci.plc"@ will display
 -- its type or an error message.
@@ -203,15 +202,16 @@ typecheckPipeline cfg =
     <=< through (unless (_tccDoNormTypes cfg) . Normal.checkProgram)
     <=< through VR.checkProgram
 
-formatDoc :: (AsParseError e AlexPosn, MonadError e m) => PrettyConfigPlc -> BSL.ByteString -> m (Doc a)
--- don't use parseScoped since we don't bother running sanity checks when we format
-formatDoc cfg = runQuoteT . fmap (prettyBy cfg) . (rename <=< parseProgram)
+formatDoc
+    :: (AsParseError e AlexPosn,
+        AsUniqueError e AlexPosn,
+        MonadError e m)
+    => PrettyConfigPlc -> BSL.ByteString -> m (Doc a)
+formatDoc cfg = runQuoteT . fmap (prettyBy cfg) . parseScoped
 
-format :: (AsParseError e AlexPosn, MonadError e m) => PrettyConfigPlc -> BSL.ByteString -> m T.Text
--- don't use parseScoped since we don't bother running sanity checks when we format
-format cfg = runQuoteT . fmap (prettyTextBy cfg) . (rename <=< parseProgram)
-
--- | Take one PLC program and apply it to another.
-applyProgram :: Program tyname name () -> Program tyname name () -> Program tyname name ()
--- TODO: some kind of version checking
-applyProgram (Program _ _ t1) (Program _ _ t2) = Program () (defaultVersion ()) (Apply () t1 t2)
+format
+    :: (AsParseError e AlexPosn,
+        AsUniqueError e AlexPosn,
+        MonadError e m)
+    => PrettyConfigPlc -> BSL.ByteString -> m T.Text
+format cfg = runQuoteT . fmap (prettyTextBy cfg) . parseScoped
