+++ language-plutus-core/src/Language/PlutusCore/Constant/Dynamic/Call.hs
@@ -9,7 +9,6 @@ module Language.PlutusCore.Constant.Dynamic.Call
 
 import           Language.PlutusCore.Constant.Make
 import           Language.PlutusCore.Constant.Typed
-import           Language.PlutusCore.Lexer.Type     hiding (name)
 import           Language.PlutusCore.Type
 
 import           Data.Proxy
