+++ language-plutus-core/src/Language/PlutusCore/Constant/Dynamic/OffChain.hs
@@ -1,11 +1,14 @@
-module Language.PlutusCore.Constant.Dynamic.OffChain ( getStringBuiltinTypes ) where
+module Language.PlutusCore.Constant.Dynamic.OffChain
+    ( getStringBuiltinTypes
+    ) where
 
-import           Control.Monad.Except
 import           Language.PlutusCore.Constant.Dynamic.BuiltinName
 import           Language.PlutusCore.Constant.Function
 import           Language.PlutusCore.Error
 import           Language.PlutusCore.Quote
 import           Language.PlutusCore.TypeCheck
+
+import           Control.Monad.Except
 
 getStringBuiltinTypes
     :: (AsTypeError e ann, MonadError e m, MonadQuote m) => ann -> m DynamicBuiltinNameTypes
@@ -14,4 +17,3 @@ getStringBuiltinTypes ann =
        insertDynamicBuiltinNameDefinition dynamicTraceDefinitionMock $
        insertDynamicBuiltinNameDefinition dynamicCharToStringDefinition $
        insertDynamicBuiltinNameDefinition dynamicAppendDefinition mempty
-
