+++ language-plutus-core/src/Language/PlutusCore/DeBruijn.hs
@@ -18,17 +18,16 @@ module Language.PlutusCore.DeBruijn
     ) where
 
 import           Language.PlutusCore.Name
-import           Language.PlutusCore.Pretty
 import           Language.PlutusCore.Quote
 import           Language.PlutusCore.Type
 
 import           Control.Exception
-import           Control.Lens               hiding (Index, Level, index, ix)
+import           Control.Lens              hiding (Index, Level, index, ix)
 import           Control.Monad.Except
 import           Control.Monad.Reader
 
-import qualified Data.Bimap                 as BM
-import qualified Data.Text                  as T
+import qualified Data.Bimap                as BM
+import qualified Data.Text                 as T
 import           Data.Typeable
 
 import           Numeric.Natural
