+++ language-plutus-core/src/Language/PlutusCore/Type/Instance.hs
@@ -1,0 +1,4 @@
+module Language.PlutusCore.Type.Instance (module Export) where
+
+import           Language.PlutusCore.Type.Instance.Pretty    as Export
+import           Language.PlutusCore.Type.Instance.Recursive as Export
